IGNITE-23562 SQL Calcite: Index count scan as new node
