+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/exec/IndexScan.java
@@ -58,7 +58,7 @@ import org.jetbrains.annotations.Nullable;
 /**
  * Scan on index.
  */
-public class IndexScan<Row> extends AbstractCacheScan<Row> {
+public class IndexScan<Row> extends AbstractCacheColumnsScan<Row> {
     /** */
     private final GridKernalContext kctx;
