+++ modules/calcite/src/main/java/org/apache/ignite/internal/processors/query/calcite/schema/IgniteIndex.java
@@ -88,9 +88,9 @@ public interface IgniteIndex {
      * @param ectx Execution context.
      * @param grp  Colocation group.
      * @param notNull Exclude null values.
-     * @return Index records number for {@code group}.
+     * @return Iterable with one row and one col: records count by index for {@code group}.
      */
-    public long count(ExecutionContext<?> ectx, ColocationGroup grp, boolean notNull);
+    public <Row> Iterable<Row> count(ExecutionContext<Row> ectx, ColocationGroup grp, boolean notNull);
 
     /**
      * Takes only first or last not-null index value.
